@@ -1,7 +1,7 @@
-package com.mythicalcreaturesllc.breedingandanimaltracking
+package com.mythcreatures.breedingtracker
 
-import android.support.test.InstrumentationRegistry
-import android.support.test.runner.AndroidJUnit4
+import androidx.test.platform.app.InstrumentationRegistry
+import androidx.test.ext.junit.runners.AndroidJUnit4
 
 import org.junit.Test
 import org.junit.runner.RunWith
@@ -19,6 +19,6 @@ class ExampleInstrumentedTest {
     fun useAppContext() {
         // Context of the app under test.
         val appContext = InstrumentationRegistry.getTargetContext()
-        assertEquals("com.mythicalcreaturesllc.breedingandanimaltracking", appContext.packageName)
+        assertEquals("com.mythcreatures.breedingtracker", appContext.packageName)
     }
 }
